% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_bibs}
\alias{download_bibs}
\title{Download .bib files}
\usage{
download_bibs(to = getwd())
}
\arguments{
\item{to}{A file path to a directory for the downloaded .bib files.}
}
\description{
Download .bib files
}
\author{
Kelli F. Johnson
}
