% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bibs.r
\name{find_bibs}
\alias{find_bibs}
\title{Find and list the names of available .bib files in \pkg{journals}}
\usage{
find_bibs()
}
\description{
Search the GitHub repository in the inst/bib directory for .bib files and
return a list of all the available files. This list can then be used within
a markdown file to specify the available .bib files.
}
\author{
Kelli F. Johnson
}
